+++ test/data/example-python-package/src/example_python_package/__init__.py
@@ -1,0 +1,2 @@
+def hello(who):
+    print(f"Hello, {who}!")